MAINT: Move preconditioners and factory functions to separate file
